backend/requirements.txt
fastapi  # Framework principal para APIs
uvicorn  # Servidor ASGI recomendado para FastAPI
pandas  # Procesamiento y análisis de datos
python-multipart  # Necesario para manejar archivos en FastAPI
aiofiles  # Requerido para manejo de archivos asíncrono
pydantic  # Validación de datos, incluido en fastapi
python-dotenv  # Variables de entorno desde .env
openai  # SDK para llamadas a OpenAI GPT (opcional, o sustituir por otro LLM)
